move to modules dir, disable animation
